refactor: share buildPagination and parseSort via query-options module
cli.js and commands/research.js previously held separate copies with
diverging semantics. Both now import one canonical implementation that
validates --limit as a positive integer; cli.js re-exports for existing
importers.
