+++ .changeset/strict-limit-validation.md
@@ -1,0 +1,5 @@
+---
+"nansen-cli": patch
+---
+
+Reject non-integer or non-positive `--limit` values with an actionable error instead of forwarding them to the API.
